Add additional runtime validation to terraform_output.
