+++ plugins/module_utils/utils.py
@@ -2,11 +2,14 @@ import os
 import json
 from typing import List, Optional, Union, cast
 
+from ansible.module_utils.compat.version import LooseVersion
+
 from ansible_collections.cloud.terraform.plugins.module_utils.types import (
     TJsonBareValue,
     TJsonObject,
     AnsibleRunCommandType,
 )
+from ansible_collections.cloud.terraform.plugins.module_utils.terraform_commands import TerraformCommands
 from ansible_collections.cloud.terraform.plugins.module_utils.errors import TerraformWarning, TerraformError
 
 
@@ -48,3 +51,36 @@ def get_outputs(
         else:
             outputs = cast(TJsonObject, json.loads(outputs_text))
             return outputs
+
+
+def validate_project_path(project_path: str) -> None:
+    if project_path is None or "/" not in project_path:
+        raise TerraformError("Path for Terraform project can not be None or ''.")
+
+    if not os.path.isdir(project_path):
+        raise TerraformError(
+            "Path for Terraform project '{0}' doesn't exist on this host - check the path and try again please.".format(
+                project_path
+            )
+        )
+
+
+def validate_bin_path(bin_path: str) -> None:
+    if not os.path.exists(bin_path):
+        raise TerraformError(
+            "Path for Terraform binary '{0}' doesn't exist on this host - check the path and try again please.".format(
+                bin_path
+            )
+        )
+
+
+def preflight_validation(
+    terraform: TerraformCommands,
+    bin_path: str,
+    project_path: str,
+    version: LooseVersion,
+    variables_args: List[str],
+) -> None:
+    validate_project_path(project_path)
+    validate_bin_path(bin_path)
+    terraform.validate(version, variables_args)
